pyproject.toml
[tool.poetry]
authors = ["gitfilin <vy.filin@yandex.ru>"]
description = ""
name = "hexlet-code"
packages = [{ include = "gendiff" }]
readme = "README.md"
version = "0.1.0"

[tool.poetry.dependencies]
python = "^3.10"
pyyaml = "^6.0.1"

[tool.poetry.dev-dependencies]
flake8 = "^7.0.0"
pytest = "^8.2.1"
pytest-cov = "^5.0.0"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]

[tool.pytest.ini_options]
addopts = "--cov=gendiff --cov-report=term-missing"
